requirements.txt
alembic==1.12.0
anyio==4.0.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
python-dotenv==1.0.0
python-telegram-bot==13.7
pytz==2023.3.post1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
